+++ doc/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 Changes in Version 4.1
 ----------------------
 
+PyMongo 4.0 brings a number of improvements including:
+
 - :meth:`pymongo.collection.Collection.update_one`,
   :meth:`pymongo.collection.Collection.update_many`,
   :meth:`pymongo.collection.Collection.delete_one`,
@@ -15,6 +17,10 @@ Changes in Version 4.1
   and :meth:`pymongo.collection.Collection.find` all support a new keyword
   argument ``let`` which is a map of parameter names and values. Parameters
   can then be accessed as variables in an aggregate expression context.
+- :meth:`~pymongo.collection.Collection.aggregate` now supports
+  $merge and $out executing on secondaries on MongoDB >=5.0.
+  aggregate() now always obeys the collection's :attr:`read_preference` on
+  MongoDB >= 5.0.
 
 
 Changes in Version 4.0